fix build
M    external/easylzma/build.rb


git-svn-id: svn+ssh://corpsvnproxy/yahoo/BrowserPlus/public_platform/trunk@475 3fcf768e-b076-0410-b9b8-e8c34e2d470b

+++ external/easylzma/build.rb
@@ -44,7 +44,7 @@ unpack(tarball)
 
 # since tarball came from github, it unpacks into a whacky name
 whackyName = Dir.glob("lloyd-easylzma-*")[0]
-if File.exists?(whackyName)
+if whackyName && File.exists?(whackyName)
     FileUtils.rm_rf(pkgDir)
     FileUtils.mv(whackyName, pkgDir)
 end
